requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
assemblyai>=0.26.0
openai>=1.12.0
pydub>=0.25.1
python-dotenv>=1.0.0
